refactor(vault): move ui logic and models to presentation layer
