+++ composeApp/src/commonMain/kotlin/org/thejohnsondev/vault/di/KoinInitializer.kt
@@ -4,7 +4,7 @@ import com.thejohnsondev.data.di.authDataModule
 import com.thejohnsondev.database.di.databaseModule
 import com.thejohnsondev.datastore.di.datastoreModule
 import com.thejohnsondev.domain.di.authDomainModule
-import com.thejohnsondev.domain.models.vaultDomainModule
+import com.thejohnsondev.domain.di.vaultDomainModule
 import com.thejohnsondev.network.di.networkModule
 import com.thejohnsondev.presentation.di.authPresentationModule
 import com.thejohnsondev.presentation.di.settingsPresentationModule
