+++ core/model/src/commonMain/kotlin/com/thejohnsondev/model/vault/PasswordDto.kt
@@ -13,4 +13,4 @@ data class PasswordDto(
     override val modifiedTimeStamp: String? = null,
     override val isFavorite: Boolean = false, // Ignore on the server side
     override val categoryId: String = DEFAULT_CATEGORY_ID,
-) : VaultItem
+) : VaultItemDto
