+++ feature/vault/domain/src/commonMain/kotlin/com/thejohnsondev/domain/models/PasswordUIModel.kt
@@ -1,4 +1,7 @@
-package com.thejohnsondev.model.vault
+package com.thejohnsondev.domain.models
+
+import com.thejohnsondev.model.vault.AdditionalFieldModel
+import com.thejohnsondev.model.vault.CategoryModel
 
 data class PasswordUIModel(
     val id: String?,
