+++ feature/vault/presentation/src/commonMain/composeResources/values/strings.xml
@@ -2,4 +2,10 @@
 <resources>
     <string name="add">Add</string>
     <string name="vault">Vault</string>
+    <string name="edit">Edit</string>
+    <string name="delete">Delete</string>
+
+    <string name="created">Created: </string>
+    <string name="modified">Modified: </string>
+    <string name="more_info">More info</string>
 </resources>
